chore: Update createChatRoom

+++ src/main/java/net/schoolvery/schoolveryserver/domain/chat/repository/ChatRepository.java
@@ -1,6 +1,10 @@
 package net.schoolvery.schoolveryserver.domain.chat.repository;
 import net.schoolvery.schoolveryserver.domain.chat.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface ChatRepository extends JpaRepository <Room,Integer> {
+import java.util.UUID;
+
+public interface ChatRepository extends JpaRepository <Room,Long> {
 }
